requirements.txt
# 문서 처리
python-docx==0.8.11
docxcompose==1.4.0
mammoth==1.6.0

# HTML/웹 처리
beautifulsoup4==4.12.2
lxml>=4.9.0
playwright>=1.35.0

# AI/번역
openai>=1.0.0
anthropic>=0.7.0

# 유틸리티
python-dotenv==1.0.0
